// src/models/PackXdslContactInfos.ts
/* WARNING: This file is auto-generated . Do not edit manually. */

import { NichandleCountryEnum } from './NichandleCountryEnum';

export type PackXdslContactInfos = {
  address: string | null;
  city: string | null;
  country: NichandleCountryEnum;
  email: string;
  firstname: string | null;
  name: string | null;
  organisation: string | null;
  phone: string | null;
  phoneCountry: NichandleCountryEnum | null;
  zip: string | null;
};