// src/models/PackXdslPackAdsl.ts
/* WARNING: This file is auto-generated . Do not edit manually. */

import { PackXdslPackCapabilities } from './PackXdslPackCapabilities';
import { OrderPrice } from './OrderPrice';

export type PackXdslPackAdsl = {
  capabilities: PackXdslPackCapabilities;
  description: string | null;
  offerDescription: string;
  offerPrice: OrderPrice;
  packName: string;
};